Move ProfileContext back to pages/Profile and fix errors
