+++ src/pages/Profile/Body/Body.tsx
@@ -1,4 +1,3 @@
-import { useProfileContext } from "contexts/ProfileContext";
 import BookmarkBtn from "components/BookmarkBtn";
 import Breadcrumbs from "components/Breadcrumbs";
 import ExtLink from "components/ExtLink";
@@ -6,6 +5,7 @@ import Icon from "components/Icon";
 import Seo from "components/Seo";
 import { APP_NAME, DAPP_DOMAIN } from "constants/common";
 import { appRoutes } from "constants/routes";
+import { useProfileContext } from "../ProfileContext";
 import DonateButton from "./DonateButton";
 import GeneralInfo from "./GeneralInfo";
