+++ src/pages/Profile/Body/GeneralInfo/Content/Content.tsx
@@ -1,4 +1,4 @@
-import { useProfileContext } from "contexts/ProfileContext";
+import { useProfileContext } from "pages/Profile/ProfileContext";
 import RichText from "components/RichText";
 import Container from "./Container";
